chore: cleanup internal decorator / plugin API
