+++ core/src/commonMain/kotlin/pro/respawn/flowmvi/api/StoreConfiguration.kt
@@ -23,6 +23,6 @@ public data class StoreConfiguration<S : MVIState>(
     val coroutineContext: CoroutineContext,
     val logger: StoreLogger,
     val atomicStateUpdates: Boolean,
-    val verifyDecorators: Boolean,
+    val verifyPlugins: Boolean,
     val name: String?,
 )
