+++ core/src/jvmTest/kotlin/pro/respawn/flowmvi/test/store/UndeliveredIntentTest.kt
@@ -1,7 +1,6 @@
 package pro.respawn.flowmvi.test.store
 
 import io.kotest.core.spec.style.FreeSpec
-import io.kotest.matchers.ints.shouldBeGreaterThan
 import io.kotest.matchers.shouldBe
 import kotlinx.coroutines.CompletableDeferred
 import pro.respawn.flowmvi.dsl.intent
